feat: Included immer
Added immer to serials
Used it in place of cloneDeep
I'm not actually certain that the use of cloneDeep is even that much of an issue here, although _certainly_ it isn't being handled correctly.

With cloneDeep we already are not working with the original object, so there should be no need to then spread the cloned object etc etc, it can instead be mutated directly.

Immer enforces proper immutability, if we used it everywhere for state transformations.

I'm not 100% convinced that it's the right thing to do though, hence draft PR

Literature: https://immerjs.github.io/immer/